Added header comments to all non-Main classes and all public functions not in Main
