+++ src/main/java/Main.java
@@ -1,8 +1,6 @@
 import enums.DialogueTypes;
 import managers.InputManager;
 import managers.OutputDialogueManager;
-
-import java.util.Scanner;
 public class Main {
     private InputManager inputManager;
     private OutputDialogueManager display;
@@ -11,6 +9,10 @@ public class Main {
         display = new OutputDialogueManager();
     }
 
+    /**
+     * Calls the InputManager to process the input repeatedly until an ExitCommand
+     * is created.
+     */
     public void run() {
         display.printInteraction(DialogueTypes.GREETINGS);
         boolean isDone = false;
