+++ src/main/java/commands/DeleteCommand.java
@@ -7,13 +7,19 @@ import managers.OutputDialogueManager;
 import managers.SaveManager;
 import managers.TaskManager;
 import tasks.Task;
-
-import java.io.IOException;
-
 import static tasks.Task.getItemCount;
 
+/**
+ * Represents a command that intends to delete an item from the list of Tasks.
+ */
 public class DeleteCommand extends Command {
     private int idToDelete;
+
+    /**
+     * Constructs a DeleteCommand Object with the index of the Task to be deleted.
+     *
+     * @param idToDelete A integer representing the index of the Task that is going to be deleted.
+     */
     public DeleteCommand(int idToDelete) {
         this.idToDelete = idToDelete;
 
@@ -32,7 +38,7 @@ public class DeleteCommand extends Command {
 
     }
 
-    private static void printDeleteResults(OutputDialogueManager display, Task toBeDeleted) {
+    private void printDeleteResults(OutputDialogueManager display, Task toBeDeleted) {
         display.printInteraction(DialogueTypes.DELETE_TASK);
         toBeDeleted.printTaskWithoutId();
         display.printInteraction(DialogueTypes.COUNT_OF_TASKS);
