+++ src/main/java/commands/FindCommand.java
@@ -6,9 +6,17 @@ import managers.OutputDialogueManager;
 import managers.SaveManager;
 import managers.TaskManager;
 
+/**
+ * Represents a command to find Tasks with names that contains a specified String.
+ */
 public class FindCommand extends Command {
     private String toFind;
 
+    /**
+     * Constructs a FindCommand Object from the String containing the word or phrase to find.
+     *
+     * @param toFind A String containing the word or phrase for this command to find.
+     */
     public FindCommand(String toFind) {
         this.toFind = toFind;
     }
