+++ src/main/java/commands/HelpCommand.java
@@ -5,6 +5,9 @@ import managers.OutputDialogueManager;
 import managers.SaveManager;
 import managers.TaskManager;
 
+/**
+ * Represent a command to call and display the help menu.
+ */
 public class HelpCommand extends Command {
     @Override
     public void execute(TaskManager taskList, SaveManager storage, OutputDialogueManager display) {
