+++ src/main/java/tasks/ToDo.java
@@ -1,13 +1,20 @@
 package tasks;
 
-import enums.DialogueTypes;
-import managers.OutputDialogueManager;
+/**
+ * Represents a task with no dates involved. A <code>Todo</code> object correspond
+ * to a Task that only have a name.
+ */
 public class ToDo extends Task {
 
     public static final String TODO_CLASS = "T";
     public static final String DONE_TODO_INDICATOR = "[T] [X]";
     public static final String NOT_DONE_TODO_INDICATOR = "[T] [ ]";
 
+    /**
+     * Construct a Todo Object using the name of the object.
+     *
+     * @param itemName A String containing the name of the Todo Object.
+     */
     public ToDo(String itemName) {
         super(itemName);
         incrementItemCount();
